tests/wpt/web-platform-tests/trusted-types/location-replace.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  async_test(t => {
    createURL_policy(window)
        .then(t.step_func_done(p => {
          let url = p.createURL(location.href + "#xxx");
          location.replace(url);
          assert_equals("" + url, location.href, "location href");
    }));
  }, "location.replace via policy (successful URL transformation).");
</script>
